Assign a unique id to each compound filter and each filter
We were assigning an id corresponding to the index of the filter /
compound filter in their array, which meant that whenever an item was
removed from the array, every child under the component corresponding
to that item was incorrectly reconciled. While this behavior only causes
performance issues when using pure components, it introduces bugs as
soon as we use stateful components.